+++ R/ci-methods.R
@@ -12,7 +12,7 @@
 #' ssd_ci_methods()
 ssd_ci_methods <- function() {
   sort(c(
-    "MACL", "GMACL", "MAW1", "MAW2", "GMAW1", "GMAW2",
+    "MACL", "GMACL",
     "multi_fixed", "multi_free", "arithmetic_samples", "geometric_samples", "weighted_samples"
   ))
   ## TODO: add "MATA" method?