+++ R/gof.R
@@ -49,7 +49,7 @@
 #' @export
 #' @examples
 #' fits <- ssd_fit_dists(ssddata::ccme_boron)
-#' ssd_gof(fits)
+#' ssd_gof(fits, wt = TRUE)
 ssd_gof <- function(x, ...) {
   UseMethod("ssd_gof")
 }