+++ R/params.R
@@ -21,10 +21,10 @@
 #' @param all_dists A flag specifying whether all the named distributions must fit successfully.
 #' @param all_estimates A flag specifying whether to calculate estimates for all implemented distributions.
 #' @param at_boundary_ok A flag specifying whether a model with one or more
-#' parameters at the boundary should be considered to have converged (default = FALSE).
+#' parameters at the boundary should be considered to have converged (default = TRUE).
 #' @param average A flag specifying whether to provide model averaged values as opposed to a value for each distribution.
 #' @param bcanz A flag or NULL specifying whether to only include distributions in the set that is approved by BC, Canada, Australia and New Zealand for official guidelines.
-#' @param big.mark A string specifying used between every 3 digits to separate thousands on the x-axis.
+#' @param big.mark A string specifying the thousands separator.
 #' @param breaks A character vector
 #' @param bounds A named non-negative numeric vector of the left and right bounds for
 #' uncensored missing (0 and Inf) data in terms of the orders of magnitude
@@ -33,33 +33,29 @@
 #' @param ci A flag specifying whether to estimate confidence intervals (by bootstrapping).
 #' @param ci_method A string specifying which method to use for estimating
 #' the standard error and confidence limits from the bootstrap samples.
-#' Possible values include `ci_method = "multi_fixed"` and  `ci_method = "multi_free"`
-#' which generate the bootstrap samples using the model-averaged cumulative distribution function
-#' but differ in whether the model weights are fixed at the values for the original dataset
-#' or re-estimated for each bootstrap sample dataset.
-#' The value `ci_method = "weighted_samples"` takes bootstrap samples
+#' The default and recommended value is still `ci_method = "weighted_samples"` 
+#' which takes bootstrap samples
 #' from each distribution proportional to its AICc based weights and
 #' calculates the confidence limits (and SE) from this single set.
-#' The value `ci_method = "MACL"` (was `ci_method = "weighted_arithmetic"` but 
-#' has been soft-deprecated) which is only included for
-#' historical reasons takes the weighted arithmetic mean of the confidence
-#' limits and `ci_method = MGCL` which was included for a research paper
-#' takes the weighted geometric mean of the confidence limits.
-#' The values `ci_method = "MAW1"` and `ci_method = "MAW2"`
-#' use the two alternative equations of Burnham and Anderson to 
-#' model average the weighted standard errors and then calculate the confidence
-#' limits using the Wald approach. 
-#' Finally `ci_method = "arithmetic"` and `ci_method = "geometric"`
+#' `ci_method = "multi_fixed"` and  `ci_method = "multi_free"`
+#' generate the bootstrap samples using the model-averaged cumulative distribution function
+#' but differ in whether the model weights are fixed at the values for the original dataset
+#' or re-estimated for each bootstrap sample dataset.
+#' The value `ci_method = "MACL"` (was `ci_method = "weighted_arithmetic"`), which is only included for
+#' historical reasons, takes the weighted arithmetic mean of the confidence
+#' limits while `ci_method = GMACL` which
+#' takes the weighted geometric mean of the confidence limits was added for completeness but is also not recommended.
+#' Finally `ci_method = "arithmetic_samples"` and `ci_method = "geometric_samples"`
 #' take the weighted arithmetic or geometric mean of the values for 
 #' each bootstrap iteration across all the distributions and then
 #' calculate the confidence limits (and SE) from the single set of samples.
-#' limits.
 #' @param censoring A numeric vector of the left and right censoring values.
 #' @param color A string of the column in data for the color aesthetic.
 #' @param computable A flag specifying whether to only return fits with numerically computable standard errors.
 #' @param conc A numeric vector of concentrations to calculate the hazard proportions for.
 #' @param control A list of control parameters passed to [`stats::optim()`].
 #' @param data A data frame.
+#' @param decimal.mark A string specifying the numeric decimal point.
 #' @param delta A non-negative number specifying the maximum absolute AIC difference cutoff.
 #' Distributions with an absolute AIC difference greater than delta are excluded from the calculations.
 #' @param digits A whole number specifying the number of significant figures.
@@ -124,7 +120,7 @@
 #' @param range_shape2 A numeric vector of length two of the lower and upper bounds for the shape2 parameter.
 #' @param range_shape2 shape2 parameter.
 #' @param reweight A flag specifying whether to reweight weights by dividing by the largest weight.
-#' @param rescale A string specifying whether to leave the values unchanged ("no") or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values ("geomean") or to logistically transform ("odds").
+#' @param rescale A flag specifying whether to leave the values unchanged (FALSE) or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values (TRUE) or a string specifying whether to leave the values unchanged ("no") or to rescale concentration values by dividing by the geometric mean of the minimum and maximum positive finite values ("geomean") or to logistically transform ("odds").
 #' @param ribbon A flag indicating whether to plot the confidence interval as a grey ribbon as opposed to green solid lines.
 #' @param right A string of the column in data with the right concentration values.
 #' @param save_to NULL or a string specifying a directory to save where the bootstrap datasets and parameter estimates (when successfully converged) to.